chore: move components out of and delete redesigned folder
