+++ src/authorizations/components/AllAccessTokenOverlay.tsx
@@ -1,5 +1,6 @@
-import React, {PureComponent, ChangeEvent} from 'react'
-import {connect, ConnectedProps} from 'react-redux'
+// Libraries
+import React, {ChangeEvent, FC, useState, useContext} from 'react'
+import {useDispatch, useSelector} from 'react-redux'
 
 // Components
 import {
@@ -19,122 +20,101 @@ import {
 
 // Actions
 import {createAuthorization} from 'src/authorizations/actions/thunks'
+import {showOverlay, dismissOverlay} from 'src/overlays/actions/overlays'
+
+// Contexts
+import {OverlayContext} from 'src/overlays/components/OverlayController'
 
 // Utils
 import {allAccessPermissions} from 'src/authorizations/utils/permissions'
+import {event} from 'src/cloud/utils/reporting'
+
+// Selectors
 import {getOrg} from 'src/organizations/selectors'
 import {getMe} from 'src/me/selectors'
 
-// Decorators
-import {ErrorHandling} from 'src/shared/decorators/errors'
-
 // Types
-import {AppState, Authorization} from 'src/types'
+import {Authorization} from 'src/types'
 
 interface OwnProps {
   onClose: () => void
 }
 
-interface State {
-  description: string
-}
+const AllAccessTokenOverlay: FC<OwnProps> = props => {
+  const {onClose} = useContext(OverlayContext)
+  const dispatch = useDispatch()
+  const [description, setDescription] = useState<string>('')
+  const {id: orgID} = useSelector(getOrg)
+  const {id: meID} = useSelector(getMe)
 
-type ReduxProps = ConnectedProps<typeof connector>
-type Props = OwnProps & ReduxProps
+  const handleSave = () => {
+    const token: Authorization = {
+      orgID,
+      description,
+      permissions: allAccessPermissions(orgID, meID),
+    }
+    dispatch(createAuthorization(token))
+    handleDismiss()
+    event('token.allAccess.create.success', {meID, name: description})
+    dispatch(showOverlay('access-token', null, () => dismissOverlay()))
+  }
 
-@ErrorHandling
-class AllAccessTokenOverlay extends PureComponent<Props, State> {
-  public state = {description: ''}
+  const handleInputChange = (event: ChangeEvent<HTMLInputElement>) => {
+    const {value} = event.target
+    setDescription(value)
+  }
 
-  render() {
-    const {description} = this.state
+  const handleDismiss = () => {
+    props.onClose()
+  }
 
-    return (
-      <Overlay.Container maxWidth={500}>
-        <Overlay.Header
-          title="Generate All Access API Token"
-          onDismiss={this.handleDismiss}
-        />
-        <Form onSubmit={this.handleSave}>
-          <Overlay.Body>
-            <FlexBox
-              alignItems={AlignItems.Center}
-              direction={FlexDirection.Column}
-              margin={ComponentSize.Large}
-            >
-              <Alert
-                icon={IconFont.AlertTriangle_New}
-                color={ComponentColor.Warning}
-              >
-                This token will be able to create, update, delete, read, and
-                write to anything in this organization
-              </Alert>
-              <Form.Element label="Description">
-                <Input
-                  placeholder="Describe this new API token"
-                  value={description}
-                  onChange={this.handleInputChange}
-                  testID="all-access-token-input"
-                />
-              </Form.Element>
-            </FlexBox>
-          </Overlay.Body>
+  return (
+    <Overlay.Container maxWidth={500}>
+      <Overlay.Header
+        title="Generate All Access API Token"
+        onDismiss={onClose}
+      />
+      <Overlay.Body>
+        <Form onSubmit={handleSave}>
+          <FlexBox
+            alignItems={AlignItems.Center}
+            direction={FlexDirection.Column}
+            margin={ComponentSize.Large}
+          >
+            <Alert icon={IconFont.AlertTriangle} color={ComponentColor.Warning}>
+              This token will be able to create, update, delete, read, and write
+              to anything in this organization
+            </Alert>
+            <Form.Element label="Description">
+              <Input
+                placeholder="Describe this new API token"
+                value={description}
+                onChange={handleInputChange}
+                testID="all-access-token-input"
+              />
+            </Form.Element>
 
-          <Form.Footer>
-            <Overlay.Footer>
+            <Form.Footer>
               <Button
                 text="Cancel"
                 color={ComponentColor.Tertiary}
-                onClick={this.handleDismiss}
+                icon={IconFont.Remove}
+                onClick={handleDismiss}
               />
+
               <Button
                 text="Save"
                 testID="button--save"
+                icon={IconFont.Checkmark}
                 color={ComponentColor.Success}
                 type={ButtonType.Submit}
               />
-            </Overlay.Footer>
-          </Form.Footer>
+            </Form.Footer>
+          </FlexBox>
         </Form>
-      </Overlay.Container>
-    )
-  }
-
-  private handleSave = () => {
-    const {orgID, meID, onCreateAuthorization} = this.props
-
-    const token: Authorization = {
-      orgID,
-      description: this.state.description,
-      permissions: allAccessPermissions(orgID, meID),
-    }
-    onCreateAuthorization(token)
-
-    this.handleDismiss()
-  }
-
-  private handleInputChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const {value} = e.target
-
-    this.setState({description: value})
-  }
-
-  private handleDismiss = () => {
-    this.props.onClose()
-  }
+      </Overlay.Body>
+    </Overlay.Container>
+  )
 }
 
-const mstp = (state: AppState) => {
-  return {
-    orgID: getOrg(state).id,
-    meID: getMe(state).id,
-  }
-}
-
-const mdtp = {
-  onCreateAuthorization: createAuthorization,
-}
-
-const connector = connect(mstp, mdtp)
-
-export default connector(AllAccessTokenOverlay)
+export default AllAccessTokenOverlay
