+++ src/authorizations/components/DisplayTokenOverlay.tsx
