+++ src/authorizations/components/IndividualAccordionBody.tsx
@@ -1,6 +1,6 @@
 // Libraries
 import React, {FC} from 'react'
-import 'src/authorizations/components/redesigned/customApiTokenOverlay.scss'
+import 'src/authorizations/components/customApiTokenOverlay.scss'
 
 // Clockface
 import {
