+++ src/authorizations/pagination/TokensTab.tsx
@@ -19,7 +19,7 @@ import SearchWidget from 'src/shared/components/search_widget/SearchWidget'
 import TokenList from 'src/authorizations/pagination/TokenList'
 import FilterList from 'src/shared/components/FilterList'
 import TabbedPageHeader from 'src/shared/components/tabbed_page/TabbedPageHeader'
-import GenerateTokenDropdownRedesigned from 'src/authorizations/components/redesigned/GenerateTokenDropdown'
+import GenerateTokenDropdown from 'src/authorizations/components/GenerateTokenDropdown'
 import ResourceSortDropdown from 'src/shared/components/resource_sort_dropdown/ResourceSortDropdown'
 import PostDeploymentTokensBanner from 'src/authorizations/components/PostDeploymentTokensBanner'
 
@@ -123,7 +123,7 @@ class TokensTab extends PureComponent<Props, State> {
       </>
     )
 
-    const rightHeaderItems = <GenerateTokenDropdownRedesigned />
+    const rightHeaderItems = <GenerateTokenDropdown />
 
     const tokensBanner = () => {
       return (
@@ -234,7 +234,7 @@ class TokensTab extends PureComponent<Props, State> {
           <EmptyState.Text>
             Looks like there aren't any <b>Tokens</b>, why not generate one?
           </EmptyState.Text>
-          <GenerateTokenDropdownRedesigned />
+          <GenerateTokenDropdown />
         </EmptyState>
       )
     }
