+++ src/overlays/components/OverlayController.tsx
@@ -15,7 +15,7 @@ import {AppState} from 'src/types'
 // Components
 import {Overlay} from '@influxdata/clockface'
 import NoteEditorOverlay from 'src/dashboards/components/NoteEditorOverlay'
-import AllAccessTokenOverlay from 'src/authorizations/components/redesigned/AllAccessTokenOverlay'
+import AllAccessTokenOverlay from 'src/authorizations/components/AllAccessTokenOverlay'
 import TelegrafConfigOverlay from 'src/telegrafs/components/TelegrafConfigOverlay'
 import TelegrafOutputOverlay from 'src/telegrafs/components/TelegrafOutputOverlay'
 import OrgSwitcherOverlay from 'src/pageLayout/components/OrgSwitcherOverlay'
@@ -33,8 +33,8 @@ import NewThresholdCheckEO from 'src/checks/components/NewThresholdCheckEO'
 import NewDeadmanCheckEO from 'src/checks/components/NewDeadmanCheckEO'
 import AutoRefreshOverlay from 'src/dashboards/components/AutoRefreshOverlay'
 import CellCloneOverlay from 'src/shared/components/cells/CellCloneOverlay'
-import CustomApiTokenOverlay from 'src/authorizations/components/redesigned/CustomApiTokenOverlay'
-import DisplayTokenOverlay from 'src/authorizations/components/redesigned/DisplayTokenOverlay'
+import CustomApiTokenOverlay from 'src/authorizations/components/CustomApiTokenOverlay'
+import DisplayTokenOverlay from 'src/authorizations/components/DisplayTokenOverlay'
 
 // Actions
 import {dismissOverlay} from 'src/overlays/actions/overlays'
